requirements.txt
langchain>=0.1.14
langchain-google-genai>=0.0.8
python-dotenv>=1.0.1
pandas>=2.0.0
